Several improvements to synchronization in notification handling.
Instead of dispatching delegate calls async to main, the notification
handler itself is async but calls delegate methods synchronously, so an
entire notification is handled, and the table view data source methods
are all called, before another notfication can be handled.

Also, for safety, at least for now, have additions, deletions, and
changes all be processed the same way, where we explicitly check if each
object exists in the before and after sets. That way, a reported deleted
isn't passed along to the delegate unless it actually disappeared. This
potentially results in a blank row, but at least doesn't crash.

The funny add/delete handling is necessary, apparently, because it
appears to be possible for a delete that has been committed to not be
reflected in a subsequent query on a different thread; I'm not sure why
that would happen.