Added a category to RLMObject to support editing an object and automatically calling the notification:
changeWithNotification:

accepts a block that itself is passed the object calling the method. This prevents forgetting to call the notification manager but given that the object within the block needs to be cast back into the subclass to use dot notation, the amount of lines for a simply change increases from 2 to 4.