API Change:
RBQRealmNotificationManager now uses a different scheme to pass changes to any listeners. Instead of passing arrays of added/deleted/inserted RBQSafeRealmObjects that are not sorted by entity type, the notification manager now passes a NSDictionary that is keyed by the class name of the entity type. The object in the dictionary is an instance of RBQEntityChangesObject which contain the added/deleted/inserted arrays for only that entity.

This improves performance when tracking changes from many entity types for different listeners. A listener simply needs to check if there is an object for the class name key it is looking for, and operate on those changes versus enumerating all of the chanegs to find the appropriate ones.

Other:
SLowly improving the internal documentation to be in line with appledoc.