API Change:
Since internally the RBQRRealmNotificationManager uses a mutable set to hold the logged changes, the notification now passes RBQEntityChangesObject that represent the add/delete/change objects in an NSSet rather than an NSArray. This makes clear all the values are unique and the changes are not returned in the order they were logged.

Other:
All headers now use appledoc style documentation formatting.