Critical fix:
FRC internals did not properly handled insertions when the current count was less than the number of items being inserted. Depending on how the processing occurred, the internal section array could have no objects in it, but an insertion would attempt to happen on an index > the array's count.

Now the insertions are sorted as they are inserted through

indexOfObject:inSortedRange:options:usingComparator:

which uses binary search to calculate the index the item should be inserted to preserve the sort.

Minor Change:

Internally the FRC now uses NSMutableOrderedSet/NSOrderedSet because there is a slight performance improvement over NSMutableArray/NSArray.