A couple more fixes to cache objects for non-string primary keys.
Also, now explicitly only handles string and int primary keys and throws an exception if you try to use a different kind. Those are the only types that Realm supports as of 0.90.5 but that could change and if Realm introduces new allowed primary key types we should fail more explicitly.