Fixed two major bugs:
1. For inserts and moves, the new cacheObject wasn't being added to the cache.Objects RLMArray when inserted (for moves, we delete the old object then insert the new one). As a result, this would affect the identification of moves, because we were comparing the count of objects in the cache to the fetch results, and using that to determine moves.

2. After realizing #1, this pointed to a bigger problem which was that our previous way of identifying moves was not accurate, and assumed that all inserts/deletes would happen at in
dex 0 (which was the case for the demo project).

The correct logic requires us to know the absolute change to the section/row for a given indexPath, taking into account the relative change from any insert or delete.

To identify the relative section changes, we create a collection of section indexes for both the deleted and inserted sections.

So for example if we insert a section at index 1, the collection would be [@(1)].

Later on, if we need to identify the count of inserted sections that would affect a given indexPath, we ask the collection for the index needed to insert the section index into the collection. Thus, if we were asking for the relative section change for a section with index 2, the result would be 1 (1 is at index 0 in the collection, so 2 would be index 1). We translate this as meaning 1 section was inserted before index 2, which is the relative section change due to inserts. This same logic is applied for deletes, and we use the difference of inserts - deletes to calculate the relative section change.

Similarly, for row changes, we create a map of objectChanges sorted by indexPath for each section. This way we can quickly retrieve the objectChanges to a given section, and then compute the relative row changes from that subset.

Just like the relative section changes, the collection of objectChanges allows us to ask for the index required to insert an objectChange. For example, if the collection had an objectChange with row [3], and we were trying to figure out if row 2 had a relative change, the result would be 0 (since we would insert 2 at index 0 for that collection). We perform this calculation for row inserts and deletes, giving us the relative row change.

With the relative section and row changes, we can decide if there was an absolute section change and report the move for that and if not, then calculate the absolute row change, and report a move if necessary for that.

If both fail, then the change is reported as an update.