Apparently UITableview's want a change reported as a move, if the object's change results in the section being deleted and a new section created, such that the indexPath is the same as before.
For example, if an object was at section: 1 row : 0 and then a change occurred which moved the object to a new section, causing the old section 1 to be deleted (there was no more items), but this new section is at index 1 too.

So the object was at section: 1 row: 0 and after the change it is still at section: 1 row: 0.... but apparently whatever is going on internally with the UITableView's batch processin
g, causes an assertion failure, since the sections are changing.

Now the processing catches this unique situation and handles it accordingly.