Performance optimization to make sure scrolling a tableview can achieve 60 fps. Calling objectAtIndexPath was slow because accessing the RLMRealm for the cache and then for the fetchRequest added ~25 ms each, and for 60 fps, the entire cell creation must take <16.67 ms.
Now the RLMRealm for the internal FRC cache and the RLMRealm for the fetchRequest is in a sense cached. A strong reference is held for each RLMRealm when an initial request is made for those properties on the main thread. Since RLMRealm's can only be accessed on the same thread, this caching only occurs for the main thread, which is also the thread in which performance is an issue.

Finally, there a write transaction needed to be called when calling reset.

Also added protection against errors when creating the cache would attempt to save an object that was already created (using addOrUpdateObject rather than addObject).