Adding back the code that tracks the state of the cache for the FRC.
This effectively reverts: https://github.com/Roobiq/RBQFetchedResultsController/commit/5122696c6d586c6e64339f13d93a1beb22689684

However, instead of a clean revert, the state change that occurs based on the Realm notification is dispatched async. This is because Realm doesn't recommend performing another write synchronously from the notification of a write, because technically the initial write isn't done yet.

This means that there is a gap between when the cache technically is ready but not yet updated, but this is acceptable, since it is more safe to assume the cache is ready when it previously failed. (Catches the scenario where the app crashes mid way through processing the cache).

Also, a change was made to the previous fix:
https://github.com/Roobiq/RBQFetchedResultsController/commit/bcbf1a762a58229fd1c0070ac35197bf05061cca

There was another exception identified to my previous fix to identify moves from updates when an object moves between section, but due to the ordering the new section's index is the same as the old one which got deleted.

The exception here was that if an item is in section 0 and row 0, with additional rows after it in that section, and then gets moved to a new section that is inserted at index 0, an update would be reported. This was because the new section is inserted, but there was no section that was deleted. My fix required the new section to be both in the inserted sections and deleted.