Fixed an issue that would occur if reset is called while a subsequent change is waiting on the semaphore to process. The issue is that the cache is deleted and rebuilt when calling reset, and once this is done the queue change would immediately happen, but the cacheRealm instance that is used to process the change hasn't actually been updated through autoRefresh due to the reset. As a result, the fix is to simply call refresh on the cacheRealm before any change processing happens.
This solves the whack-a-mole that was happening in which objects would appear as invalidated.

Finally, added protection to check if an indexPath is represented in the cache and if not return nil for objectAtIndexPath methods.

Cleaned up the log statements that were kept to help debug the issue in v1.6.1