Now if you supply nil for the cacheName, the cache will not be persisted and an in-memory Realm will be used.
Re-reading the Realm docs, noticed that in-memory Realms can be used across threads with the identifier as long as the object is held strongly. Given that the FRC now holds a strong reference to the RLMRealm for the cache on the main thread to improve scroll performance.... the side benefit is now in-memory Realm's work with nominal changes.

Still leaving the explicity constructor method to use a supplied in-memory Realm, but will reevaluate whether this is necessary.