Removing the logic that would create the cache on demand when one of the public methods such as numberOfSections was called.
This shouldnt't be the FRC's decision, but rather the methods should return expected values, such as 0 for numberOfSections, and then the developer can decide when to call perfromFetch.

This is problematic when the FRC backs a UITableview and its delegates call the FRC on the main thread, forcing the performFetch to be on the main thread rather than on a background thread called elsewhere.