Now calling refresh on the Realm passed in through the change notification in addition to refreshing the cache Realm. There have been observed mishandling of deleting sections that have no more rows in them. Haven't been able to actually reliably reproduce, which leads me to believe that this is due to not having all the changes merged into the Realm used for the change processing.
Specifically, sections that need removed are identified by asking the Realm for all objects that match the sectionNameKeyPath for the section, if the results have a count of zero, then we remove the section. This logic could break down where the count comes back greater than zero because the Realm is not up to date with changes, but later on the row level changes moves items out of that section, leaving a section with no rows.