Fixes to support Swift and nil predicates.
Swift class names are programatically referenced by "AppName.ClassName" instead of just "ClassName". The constructor methods now verify that the entity name submitted matches an actual class or attempts to build the appropriate Swift class name.

The evaluateObject: method supports a nil predicate. If there is no predicate the only evaluation is to make sure the object is of the same class name matches the RBQFetchRequest entity name.