Fix to bring back the improvement from the original commit: https://github.com/Roobiq/RBQFetchedResultsController/commit/d48d174055e66479354abab8b8ce0845de36fe32
which attempted to prevent the main RLMRealm from having stale data while doing change processing by calling refresh on it.

This however, would create a recursion since refresh itself triggers RLMRealmDidChangeNotification(s).

To fix this now there is a helper refresh method in the FRC that uses obj-c run-time to associate an object with the RLMRealm. In the notification block code, a check was added to see if the RLMRealm has the associated object and to simply return if so. This prevents the change processing from recursing if we need to call refresh.

Right now, the refresh is called at the beginning of the change processing and not elsewhere, but potentially it might need to be called in more places. However, I am unable to reproduce the original issue that started all of this at this point, so until then, I will leave it as is.