If edits were to occur when the app is in background state, turns out the directories that held the FRC's caches were not accessible. Apparently if you don't specify a file protection when creating directories through the NSFileManager method:
createDirectoryAtPath:withIntermediateDirectories:attributes:erro:

this will default to NSFileProtectionComplete. Unknown if this is a new default, since the Apple docs aren't too helpful:

"Permissions are set according to the umask of the current process. For more information, see umask."

But no where could I find what the default is. Anyway, now the directories are specified as NSFileProtectionNone since the Realm files themselves can be protected absent the protection of their directories.