Reworked the internal processing of the FRC to not block on the main thread updating. This should prevent deadlocks when the developer mixes changes occurring on background and main thread.
To do so, the FRC uses an internal serial queue and does all the processing to identify the changes synchronously with the notification call. Then it dispatches to the serial queue the code that actually adjusts the internal cache and dispatches the delegate calls back to the main thread.

Seperating out to a seperate branch for testing and further clean up.