Removing the use of tracking "state" in the cache. This is not relevant given that its purpose is to track if the cache was left in a processing state and needs rebuilt. The problem is that this cannot occur. The state is changed within a write transaction block on the cache Realm, and if the app would crash during this write transaction, it will never be committed.
The result of tracking the state meant that since the FRC internally is not using a queue to simplify the thread management, it is difficult to synchronize both of the notification blocks together with other calls to updateFetchRequest: or reset.

Not actually removing the property from the cache object (just adding deprecation notification) because this would force users to migrate.