Deprecated:
1. initWithFetchRequest:sectionNameKeyPath:inMemoryRealmCache:
2. objectInRealm:atIndexPath:

in RBQFetchedResultsController

The reason being, when the class was first created In-Memory Realms were not accessible across threads. This has since changed and it is possible to store the in-memory Id (as long as the in-memory Realm is strongly held somewhere) and grab a new instance on a different thread.

As a result, RBQFetchRequest now stores the identifier and can retrieve the in-memory Realm when calling its realm property, allowing the FRC to use this, making these methods obsolete.

Furthermore, the external in-memory Realm cache is unnecesary now that passing nil for the cache name internally uses an in-memory Realm. This mirrors NSFetchedResultsController.

Finally, added a test for in-memory Realm based RBQFetchRequest and cleaned up the recent PR formatting.