Removing use of sorting during iteration via binary search. Now sorts are applied after the iteration. Should be a minor speed up.
However, in the process found a code error that potentially was causing errors. The inserted changes for a section wasn't actually getting the object change. The object change was mistakingly being added twice to the inserted object changes set. This didn't result in a duplicate because it is a set, but the other collection was lacking it!